+++ test/lib/delay.ts
@@ -1,5 +1,5 @@
 export function delay(ms: number): Promise<void> {
-  return new Promise((resolve, _reject) => {
+  return new Promise(resolve => {
     setTimeout(resolve, ms)
   })
 }