+++ test/redis-failure.ts
@@ -3,7 +3,7 @@ import {expect} from "chai"
 import {And, Feature, Given, Scenario, Then, When} from "./lib/steps"
 
 import uuidv1 from "uuid/v1"
-import {Redis, SafeRedisSlidingWindowRateLimiter} from "../src/sliding-window-rate-limiter"
+import {CheckResult, Redis, ReserveResult, SafeRedisSlidingWindowRateLimiter} from "../src/sliding-window-rate-limiter"
 import {MockIORedis} from "./lib/mock-ioredis"
 
 Feature("Test sliding-window-rate-limiter Redis failure with safe backend", () => {
@@ -11,9 +11,11 @@ Feature("Test sliding-window-rate-limiter Redis failure with safe backend", () =
     const defaultLimit = 1
     const errors: Error[] = []
 
-    let redis: Redis
-    let limiter: SafeRedisSlidingWindowRateLimiter
+    let checkResult: CheckResult
     let key: string
+    let limiter: SafeRedisSlidingWindowRateLimiter
+    let redis: Redis
+    let reserveResult: ReserveResult
 
     Given("redis connection", () => {
       redis = new MockIORedis()
@@ -41,18 +43,26 @@ Feature("Test sliding-window-rate-limiter Redis failure with safe backend", () =
     })
 
     And("check method is called", async () => {
-      expect(await limiter.check(key, defaultLimit)).to.equal(0)
+      checkResult = await limiter.check(key, defaultLimit)
     })
 
-    Then("one error event was fired", () => {
+    Then("usage is 0", async () => {
+      expect(checkResult.usage).to.equal(0)
+    })
+
+    And("one error event was fired", () => {
       expect(errors.length).to.equal(1)
     })
 
     When("reserve method is called", async () => {
-      expect(await limiter.reserve(key, defaultLimit)).to.equal(0)
+      reserveResult = await limiter.reserve(key, defaultLimit)
     })
 
-    Then("one error event was fired", () => {
+    Then("usage is 0", async () => {
+      expect(reserveResult.usage).to.equal(0)
+    })
+
+    And("one error event was fired", () => {
       expect(errors.length).to.equal(1)
     })
   })